Cherry-picking some changes from feature/agendaOnline
