+++ TripKit/AddOns/Bookings/AccountKit/Model/Server/AMKCommunicator.m
@@ -584,12 +584,13 @@ static NSString *const kPathImage     = @"api/user/image";
   [self.sharedServer hitSkedGoWithMethod:@"POST"
                                     path:kPathImage
                               parameters:nil
+                                 headers:nil
                               customData:imageData
                                   region:nil
                           callbackOnMain:YES
                                  success:
-   ^(NSInteger status, id responseObject, NSData *data) {
-#pragma unused(status, data)
+   ^(NSInteger status, NSDictionary<NSString *, id> *headers, id responseObject, NSData *data) {
+#pragma unused(status, headers, data)
      typeof(weakSelf) strongSelf = weakSelf;
      if (! strongSelf) return;
      if (handler != nil) {
