+++ TripKit/AddOns/Share/TKShareURLProvider.swift
@@ -41,7 +41,7 @@ public protocol TKURLSavable: class, TKURLShareable {
 
       let saveURL = self.saveURL(forBase: baseSaveURL, allowLongURL: allowLongURL)
       
-      SVKServer.get(saveURL, paras: nil) { _, response, _, _ in
+      SVKServer.get(saveURL, paras: nil) { _, _, response, _, _ in
         guard
           let dict = response as? [String: Any],
           let urlString = dict["url"] as? String,
