+++ TripKit/Classes/core/ServerKit/SVKServer.h
@@ -16,8 +16,9 @@
 NS_ASSUME_NONNULL_BEGIN
 
 typedef void (^SGServerSuccessBlock)(NSInteger status, id _Nullable responseObject, NSData * _Nullable data);
+typedef void (^SGServerFullSuccessBlock)(NSInteger status, NSDictionary<NSString *, id> *headers, id _Nullable responseObject, NSData * _Nullable data);
 typedef void (^SGServerFailureBlock)(NSError *error);
-typedef void (^SGServerGenericBlock)(NSInteger status, id _Nullable responseObject, NSData * _Nullable data, NSError * _Nullable error);
+typedef void (^SGServerGenericBlock)(NSInteger status, NSDictionary<NSString *, id> *headers, id _Nullable responseObject, NSData * _Nullable data, NSError * _Nullable error);
 
 typedef NS_ENUM(NSInteger, SVKServerType) {
   SVKServerTypeProduction = 0,
@@ -87,7 +88,7 @@ FOUNDATION_EXPORT NSString *const SVKDefaultsKeyProfileEnableFlights;
 
 - (void)hitSkedGoWithMethod:(NSString *)method
                        path:(NSString *)path
-                 parameters:(nullable NSDictionary<NSString *, id>  *)parameters
+                 parameters:(nullable NSDictionary<NSString *, id> *)parameters
                      region:(nullable SVKRegion *)region
              callbackOnMain:(BOOL)callbackOnMain
                     success:(SGServerSuccessBlock)success
@@ -95,11 +96,21 @@ FOUNDATION_EXPORT NSString *const SVKDefaultsKeyProfileEnableFlights;
 
 - (void)hitSkedGoWithMethod:(NSString *)method
                        path:(NSString *)path
-                 parameters:(nullable NSDictionary<NSString *, id>  *)parameters
+                 parameters:(nullable NSDictionary<NSString *, id> *)parameters
+                    headers:(nullable NSDictionary<NSString *, NSString *> *)headers
+                     region:(nullable SVKRegion *)region
+             callbackOnMain:(BOOL)callbackOnMain
+                    success:(SGServerFullSuccessBlock)success
+                    failure:(SGServerFailureBlock)failure;
+
+- (void)hitSkedGoWithMethod:(NSString *)method
+                       path:(NSString *)path
+                 parameters:(nullable NSDictionary<NSString *, id> *)parameters
+                    headers:(nullable NSDictionary<NSString *, NSString *> *)headers
                  customData:(nullable NSData*)customData
                      region:(nullable SVKRegion *)region
              callbackOnMain:(BOOL)callbackOnMain
-                    success:(SGServerSuccessBlock)success
+                    success:(SGServerFullSuccessBlock)success
                     failure:(SGServerFailureBlock)failure;
 
 - (nullable id)initiateSyncRequestWithMethod:(NSString *)method
