+++ TripKit/Classes/server/TKDeparturesProvider.swift
@@ -48,8 +48,8 @@ extension TKDeparturesProvider {
     
     return SVKServer.shared.rx
       .hit(.POST, path: "departures.json", parameters: paras, region: region)
-      .map {
-        guard let data = $2 else { throw OutputError.noDataReturn }
+      .map { _, _, data in
+        guard let data = data else { throw OutputError.noDataReturn }
         let decoder = JSONDecoder()
         return try decoder.decode(API.Departures.self, from: data)
     }
@@ -148,8 +148,8 @@ extension TKDeparturesProvider {
     
     return SVKServer.shared.rx
       .hit(.POST, path: "departures.json", parameters: paras, region: table.startRegion)
-      .map {
-        guard let data = $2 else { throw OutputError.noDataReturn }
+      .map { _, _, data in
+        guard let data = data else { throw OutputError.noDataReturn }
         let decoder = JSONDecoder()
         return try decoder.decode(API.Departures.self, from: data)
     }
