+++ TripKit/Classes/server/TKQuickBookingHelper.swift
@@ -95,7 +95,7 @@ public enum TKQuickBookingHelper {
       return
     }
     
-    SVKServer.get(bookingsURL, paras: nil) { _, response, _, error in
+    SVKServer.get(bookingsURL, paras: nil) { _, _, response, _, error in
       guard let array = response as? [[String: Any]], !array.isEmpty else {
         completion([])
         SGKLog.warn("TKQuickBookingHelper", text: "Response isn't array.\nResponse: \(String(describing: response))\nError: \(String(describing: error))")
